backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-RESTX==1.1.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
alembic==1.12.0
email-validator==2.0.0.post2
python-dotenv==1.0.0
Pillow==10.0.1
pytest==7.4.2
pytest-cov==4.1.0
coverage==7.3.1
python-magic==0.4.27
bleach==6.0.0
markdown==3.4.4
WTForms==3.0.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
black==23.9.1
flake8==6.1.0
PyJWT==2.8.0 